prints all the elements of a dlistint_t list

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -6,7 +6,7 @@
  */
 size_t print_dlistint(const dlistint_t *h)
 {
-	int node;
+	size_t node;
 
 	if (h == NULL)
 	{
